Tweak color palette texel lookup to avoid interpolation errors
Dividing by the number of pixels in the texture puts the result right at the center of the texel,
with filter mode NEAREST that can flip it from one pixel to the next during rendering due to float
errors. This way it should always be on one 'side' of the texel without leaving it, preventing this
flipping behavior.